Expose binder and formatter builders in singleton methods
