+++ enumeratum-play/compat/src/main/scala-2.11/enumeratum/values/PlayQueryBindableValueEnum.scala
@@ -16,19 +16,7 @@ sealed trait PlayQueryBindableValueEnum[ValueType <: AnyVal, EntryType <: ValueE
   /**
    * Implicit path binder for Play's default router
    */
-  implicit lazy val queryBindable: QueryStringBindable[EntryType] = new QueryStringBindable[EntryType] {
-    def bind(key: String, params: Map[String, Seq[String]]): Option[Either[String, EntryType]] = {
-      baseQueryBindable.bind(key, params).map(_.right.flatMap { s =>
-        val maybeBound = enum.withValueOpt(s)
-        maybeBound match {
-          case Some(obj) => Right(obj)
-          case None => Left(s"Unknown value supplied for $enum '$s'")
-        }
-      })
-    }
-
-    def unbind(key: String, entry: EntryType): String = s"$key=${entry.value}"
-  }
+  implicit lazy val queryBindable: QueryStringBindable[EntryType] = UrlBinders.queryBinder(baseQueryBindable)(enum)
 }
 
 /**
